First moves towards Java5 code.
Currently I am in the process of repeating the following steps:
- let Eclipse infer generic arguments
- fix things when it breaks (and it does)
- run unit tests
- play a game or two to test
- commit

This commit contains the switch to Java 5 in the compiler settings plus one such cycle. More will follow. 

git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@2967 c5679a2a-0e0e-0410-bba2-8642b794b8c1

+++ Colossus/core/src/main/java/net/sf/colossus/xmlparser/StrategicMapLoader.java
@@ -82,7 +82,7 @@ public class StrategicMapLoader
         {
             Element exit = (Element)it.next();
             String sExitType = exit.getAttributeValue("type");
-            int iExitType = ((Integer)Constants.hexsideMap.get(sExitType))
+            int iExitType = Constants.hexsideMap.get(sExitType)
                 .intValue();
             hex.setBaseExitType(i, iExitType);
             int exitLabel = exit.getAttribute("label").getIntValue();
